Fixed sqrt issue for big integer issue by using Long data type

+++ src/leetcode/Sqrt.java
@@ -1,11 +1,9 @@
 package leetcode;
 
+
 /**
- * Created with IntelliJ IDEA.
- * User: Zhisheng
- * Date: 7/7/13
- * Time: 1:57 AM
- * To change this template use File | Settings | File Templates.
+ * Use binary search to calculate the sqrt of the integer.
+ * In order to handle overflow case, we need to use long data type instead of int
  */
 public class Sqrt {
 
@@ -14,28 +12,25 @@ public class Sqrt {
         // DO NOT write main() function
         if (x < 0) return -1;
 
-        return sqrt(0, x, x);
+        return (int)sqrt(0, x, x);
 
     }
 
-    private int sqrt(int start, int end, int target) {
+    private long sqrt(long start, long end, long target) {
         if (end == start) {
-            if (end * end == target || (end*end < target && (end + 1) * (end + 1) > target))
+            if (end * end == target || (end * end < target && (end + 1) * (end + 1) > target))
                 return end;
             else
                 return -1; // should never get here
         }
 
-        int mid = start + (end - start) / 2;
+        long mid = start + (end - start) / 2;
 
-        if (mid * mid == target || (mid*mid < target && (mid + 1) * (mid + 1) > target)) {
+        if (mid * mid == target || (mid * mid < target && (mid + 1) * (mid + 1) > target)) {
             return mid;
         } else if (mid * mid > target) {
             return sqrt(start, mid, target);
         } else
             return sqrt(mid + 1, end, target);
-
     }
-
-    //todo: need to handle big integer case
 }
